bin/scout: fixed command templates for scout_ssid ops
Added updated commands for scout_ssid templates. Newer Cisco
Aironet APs no longer accept d0 as an alias for the Dot11Radio
interface.

Started the groundwork for a test suite. Testing against older
Aironet 1142 series APs that may still be in use as well as the newer
2900/3700 series.

Added commanddebug as a mandatory option to specify, however,
it's not required to be on. commanddebug will allow users
to see each command being executed on the AP.

Signed-off-by: James McClune <jmcclune@mcclunetechnologies.net>