cardinal: finished all of the SSID operation tiles
Although there is much work to be done, all of the Cardinal
operations within the UI are now operational. Whether or not
they would work in a production environment is the next question.

As of now, Cardinal can be used to manage some of the aspects
of a Cisco access point. Access points can be added and deleted.
SSIDs can now be deployed on an individual AP & AP group level.

RADIUS, SNMP, and HTTP can be enabled/disabled. TFTP backups can
be kicked off on both an individual AP & AP group level.

I think I'm almost ready to cut a Cardinal v2.0 release :)

Signed-off-by: James McClune <jmcclune@mcclunetechnologies.net>